api/requirements.txt
Django==2.2.3
django-tastypie==0.14.2
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
